# back/openapi.yaml
...
openapi: 3.0.3
info:
  title: 파이어가드 API
  version: 0.7.0
  description: |
    CCTV 화재 감지 → 관리자 알림 → 무응답 시 119 자동 신고·승계 백엔드 API.

    ## 공통 규칙
    - 모든 필드는 `snake_case`, 날짜·시각은 ISO 8601 (`2026-08-10T14:30:00`)
    - 인증은 `Authorization: Bearer <access_token>` (로그인·헬스체크·미디어 제외)
    - AI 모델 전용 `/api/internal/*` 은 JWT 대신 `X-Internal-Key` 헤더를 쓴다
    - 에러는 어떤 상황에서도 `{ "code": "...", "message": "..." }` 형태다.
      검증 실패(400)는 어느 입력 탓인지 가리킬 수 있으면 `field` 가 하나 더 붙는다
      (`{ "code": "BAD_REQUEST", "message": "...", "field": "user_pw" }`) —
      400 은 `code` 가 전부 `BAD_REQUEST` 라, 폼 입력칸에 메시지를 붙일 때 이걸 쓴다
    - 목록 중 페이징이 있는 것은 `items / page / size / total_count / total_pages`
      봉투로 감싸 내려간다 (`/api/cctvs`, `/api/agencies` 는 페이징 없이 `items` 만)

    ## 실시간 알림 — 폴링하지 않는다 (2026-08-11 확정)
    WebSocket·SSE 뿐 아니라 **폴링도 쓰지 않는다**. 119 신고가 화재확인(`POST /api/alerts/{alert_no}/respond`)
    요청 안에서 **동기로** 끝나므로, 기다렸다가 알려줄 상태 변화가 없다.
    - 사람을 깨우는 채널은 **SMS** 다. 문자를 보고 앱을 열면 그때 목록을 한 번 조회하면 된다.
    - 남은 시간 카운트다운은 응답에 실린 `alert_deadline_at` 으로 클라이언트가 로컬 계산한다.
    - 화면 갱신 원칙: **열 때 조회 + 행동 후 갱신**.

    ## 판정·유예 기본값 (2026-08-10 변경)
    | 값 | 기본 | 뜻 |
    |---|---|---|
    | `EVENT_WINDOW_SEC` | 60초 | 관측 창. **최초 감지 시각에 고정**되며 이후 검출로 연장되지 않는다 |
    | `EVENT_THRESHOLD_FRAMES` | 30 | 창 안에서 이만큼 쌓이면 화재 확정. 미달인 채 창이 닫히면 `DISMISSED` |
    | `ALERT_DEADLINE_SEC` | 30초 | 알림 응답 유예. **이전 `ALERT_DEADLINE_MIN`(3분)에서 바뀐 값** |
    | `ESCALATION_INTERVAL_SEC` | 5초 | 무응답 스윕 주기. 신고가 최대 이만큼 늦어져 실제 유예는 30~35초 |

...
servers:
...
paths:
...
  /api/users:
    get:
      tags: [users]
      summary: 계정 목록 (ADMIN)
      security:
        - bearerAuth: []
      parameters:
        - name: user_status
          in: query
          description: 계정 상태 필터
          schema:
            $ref: "#/components/schemas/UserStatus"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/UserSummary"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
    post:
      tags: [users]
      summary: 계정 등록 (ADMIN)
      description: |
        아이디·비밀번호 작성규칙을 여기서만 검사한다.
        - `user_id`: 5~20자, 영문 소문자 시작, 소문자·숫자·`_`·`-`
        - `user_pw`: 영문+숫자+특수문자 3종 포함 8~64자.
          같은 문자 3연속·연속 문자열(1234, abcd)·키보드 배열(qwer)·아이디 포함 금지
        - `user_phone`: 하이픈 없이 숫자만 (`01012345678`)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserCreate"
      responses:
        "201":
          description: 등록 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_no:
                    type: integer
                    format: int64
        "400":
          description: 필수값 누락 또는 형식 오류 (`BAD_REQUEST`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                required:
                  summary: 필수값 누락
                  value:
                    code: BAD_REQUEST
                    message: user_name 는 필수입니다.
                    field: user_name
                user_id_rule:
                  summary: 아이디 작성규칙 위반
                  value:
                    code: BAD_REQUEST
                    message: user_id 는 5~20자의 영문 소문자·숫자·특수기호(_,-)만 사용할 수 있으며, 시작은 영문 소문자여야 합니다.
                    field: user_id
                password_rule:
                  summary: 비밀번호 작성규칙 위반
                  value:
                    code: BAD_REQUEST
                    message: 비밀번호는 영문자·숫자·특수문자를 모두 포함해 8자 이상이어야 합니다.
                    field: user_pw
                phone_format:
                  summary: 전화번호 형식 오류
                  value:
                    code: BAD_REQUEST
                    message: "user_phone 은 하이픈 없이 숫자만 입력해야 합니다. (예: 01012345678)"
                    field: user_phone
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "409":
          $ref: "#/components/responses/DuplicateUserId"
  /api/users/password:
    put:
      tags: [users]
      summary: 본인 비밀번호 변경
      description: |
        로그인한 사용자가 **자기 비밀번호만** 바꾼다. 대상 `user_no` 를 받지 않고
        토큰에서 꺼내 쓰므로, 남의 비밀번호를 바꿀 통로가 애초에 없다.

        현재 비밀번호를 함께 받아 대조한다 — 토큰만으로 통과시키면 자리를 비운
        사이 브라우저를 잡은 사람이 비밀번호를 갈아버릴 수 있다.

        새 비밀번호는 계정 등록·수정과 같은 작성규칙을 거친다.
        비밀번호를 잊어버린 경우는 이 API 가 아니라
        `/api/auth/password-reset/*` 쪽이다.

        소셜 로그인 계정은 바꿀 비밀번호가 애초에 없다 — 400 `SOCIAL_ACCOUNT` 다.
        프론트는 이 계정에 비밀번호 변경 화면 자체를 띄우지 않는 편이 낫다
        (`GET /api/auth/me` 로는 구분되지 않으니 소셜 로그인으로 들어왔다는 사실을
        프론트가 기억해 두면 된다).
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password:
                  type: string
                  description: 현재 비밀번호
                new_password:
                  type: string
                  description: 새 비밀번호 (작성규칙 적용)
      responses:
        "200":
          description: 변경 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
              example:
                message: 비밀번호가 성공적으로 변경되었습니다.
        "400":
          description: 입력 누락 · 소셜 계정 · 현재 비밀번호 불일치 · 작성규칙 위반
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                missing:
                  summary: 필수값 누락
                  value:
                    code: BAD_REQUEST
                    message: 현재 비밀번호와 새 비밀번호를 모두 입력해주세요.
                social:
                  summary: 소셜 로그인 계정 (바꿀 비밀번호가 없다)
                  value:
                    code: SOCIAL_ACCOUNT
                    message: 소셜 로그인 계정은 비밀번호를 변경할 수 없습니다.
                    field: user_pw
                wrong_current:
                  summary: 현재 비밀번호 불일치
                  value:
                    code: INVALID_CURRENT_PASSWORD
                    message: 현재 비밀번호가 일치하지 않습니다.
                password_rule:
                  summary: 새 비밀번호 작성규칙 위반
                  value:
                    code: BAD_REQUEST
                    message: 비밀번호는 영문자·숫자·특수문자를 모두 포함해 8자 이상이어야 합니다.
                    field: user_pw
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/UserNotFound"
  /api/users/{user_no}:
    put:
      tags: [users]
      summary: 계정 수정 · 정지 · 탈퇴
      description: |
        ADMIN 은 모든 계정의 모든 필드를, 일반 사용자는 **본인 계정의**
        `user_name`·`user_email`·`user_phone`·`user_gender`·`user_address`·`user_pw`
        만 수정할 수 있다. `user_role`·`user_status` 가 하나라도 섞이면 요청 전체를
        403 으로 거부한다 (부분 적용 없음).

        `user_status: "WITHDRAWN"` 을 보내면 탈퇴 일시가 함께 기록된다.
        보낸 필드만 갱신되며, 갱신할 필드가 하나도 없으면 400 이다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserNo"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: 수정 성공
          content:
            application/json:
              schema:
                type: object
                properties:
                  user_no:
                    type: integer
                    format: int64
        "400":
          description: 필수값 누락 또는 형식 오류 (`BAD_REQUEST`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              examples:
                nothing_to_update:
                  summary: 수정할 필드 없음
                  value:
                    code: BAD_REQUEST
                    message: 수정할 필드가 없습니다.
                password_rule:
                  summary: 비밀번호 작성규칙 위반
                  value:
                    code: BAD_REQUEST
                    message: 비밀번호에 연속된 문자열(1234, abcd 등)은 사용할 수 없습니다.
                    field: user_pw
                phone_format:
                  summary: 전화번호 형식 오류
                  value:
                    code: BAD_REQUEST
                    message: "user_phone 은 하이픈 없이 숫자만 입력해야 합니다. (예: 01012345678)"
                    field: user_phone
                social_password:
                  summary: 소셜 계정에 비밀번호 설정 시도
                  value:
                    code: SOCIAL_ACCOUNT
                    message: 소셜 로그인 계정에는 비밀번호를 설정할 수 없습니다.
                    field: user_pw
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenUserUpdate"
        "404":
          $ref: "#/components/responses/UserNotFound"
  /api/users/{user_no}/activities:
    get:
      tags: [users]
      summary: 활동이력 (지정 사용자)
      description: |
        해당 사용자의 활동이력을 **최신순**으로 준다. 남기는 종류는
        `ActivityType` 참고. IP·User-Agent 는 수집하지 않는다.

        권한은 계정 수정과 같은 규칙이다 — ADMIN 은 전체, 일반 사용자는 본인 것만.
        활동 이력은 감사 자료라 남의 것을 아무나 볼 수 없다.

        **본인 이력만 필요하면 `GET /api/me/activities` 를 쓸 것.** 이 경로는
        관리자가 남의 이력을 보는 화면(회원 상세)을 위한 것이다.

        회원 상세 화면의 "최근 로그인"은 `?size=1` 로 받은 첫 `LOGIN` 항목을 쓰면 된다.

        주의: 실패한 시도는 여기 쌓이지 않는다. 그리고 `LOGOUT` 은 프론트가
        `POST /api/auth/logout` 을 불러야만 생긴다 — 브라우저를 그냥 닫으면
        `LOGIN` 만 있고 짝이 되는 `LOGOUT` 이 없는 행이 남는다. 정상이다.
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserNo"
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/Size"
      responses:
        "200":
          description: 조회 성공 (이력이 없으면 `items` 가 빈 배열)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/PageMeta"
                  - type: object
                    properties:
                      items:
                        type: array
                        items:
                          $ref: "#/components/schemas/UserActivity"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          description: 남의 활동이력 조회 시도 (`FORBIDDEN`)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
              example:
                code: FORBIDDEN
                message: 본인 계정의 활동이력만 조회할 수 있습니다.
        "404":
          $ref: "#/components/responses/UserNotFound"

  # ---------------------------------------------------------------- 내 정보
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '로그인으로 받은 `access_token` 을 `Authorization: Bearer <token>` 으로 보낸다.'
    internalKey:
      type: apiKey
      in: header
      name: X-Internal-Key
      description: AI 모델 전용. 서버의 `INTERNAL_API_KEY` 환경변수 값과 대조한다.
  parameters:
    Page:
      name: page
      in: query
      description: 페이지 번호 (1부터)
      schema:
        type: integer
        minimum: 1
        default: 1
    Size:
      name: size
      in: query
      description: 페이지 크기 (최대 100)
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    UserNo:
      name: user_no
      in: path
      required: true
      schema:
        type: integer
        format: int64
...
  responses:
    Unauthorized:
      description: |
        토큰 없음(`UNAUTHORIZED`) · 만료(`TOKEN_EXPIRED`) · 손상(`INVALID_TOKEN`) ·
        폐기(`TOKEN_REVOKED`). 넷 다 로그인 화면으로 돌려보내야 하지만,
        **안내 문구는 갈라주는 편이 낫다.**

        `TOKEN_REVOKED` 는 사용자가 잘못한 게 없는데 튕기는 경우다 — 다른 기기에서
        로그아웃했거나 관리자가 계정을 정지했을 때 나온다. 이유를 안 알려주면
        사용자는 영문도 모르고 튕긴 것으로 느낀다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            missing:
              summary: 토큰 없음
              value:
                code: UNAUTHORIZED
                message: 인증 토큰이 필요합니다.
            expired:
              summary: 토큰 만료
              value:
                code: TOKEN_EXPIRED
                message: 토큰이 만료되었습니다. 다시 로그인해주세요.
            invalid:
              summary: 토큰 손상
              value:
                code: INVALID_TOKEN
                message: 유효하지 않은 토큰입니다.
            revoked:
              summary: 토큰 폐기 (로그아웃·계정 정지)
              value:
                code: TOKEN_REVOKED
                message: 로그아웃되었거나 무효화된 토큰입니다. 다시 로그인해주세요.
    Forbidden:
      description: 관리자 권한 필요 (`FORBIDDEN`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: FORBIDDEN
            message: 관리자 권한이 필요합니다.
...
    ForbiddenUserUpdate:
      description: |
        남의 계정을 수정하려 했거나(`FORBIDDEN`), 일반 사용자가 관리자 전용 필드
        (`user_role`·`user_status`)를 건드리려 한 경우. 후자는 요청 전체가 거부된다.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          examples:
            not_mine:
              summary: 남의 계정
              value:
                code: FORBIDDEN
                message: 본인 계정만 수정할 수 있습니다.
            admin_only_field:
              summary: 관리자 전용 필드
              value:
                code: FORBIDDEN
                message: 권한 등급과 계정 상태는 관리자만 변경할 수 있습니다.
    DuplicateUserId:
      description: 이미 사용 중인 아이디 (`DUPLICATE_USER_ID`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: DUPLICATE_USER_ID
            message: 이미 사용 중인 아이디입니다.
    UserNotFound:
      description: 사용자 없음 (`USER_NOT_FOUND`)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            code: USER_NOT_FOUND
            message: 사용자를 찾을 수 없습니다.
...
  schemas:
...
    Error:
      type: object
      description: |
        모든 에러 응답의 공통 형식.

        `field` 는 검증 실패(400)에서 **어느 입력 때문인지** 가리킬 수 있을 때만
        실린다. 400 은 `code` 가 전부 `BAD_REQUEST` 하나라, 이게 없으면 프론트가
        폼의 어느 칸에 메시지를 붙일지 정하려고 한글 문구를 파싱해야 한다.
        입력 하나를 특정할 수 없는 400(예: "수정할 필드가 없습니다")과
        400 이 아닌 에러에는 키 자체가 없다.
      required: [code, message]
      properties:
        code:
          type: string
          example: BAD_REQUEST
        message:
          type: string
          description: 사용자에게 그대로 보여줄 수 있는 한글 메시지
          example: user_id 는 필수입니다.
        field:
          type: string
          description: 문제가 된 요청 필드·쿼리 파라미터 이름 (없을 수 있음)
          example: user_id
    PageMeta:
      type: object
      description: 페이징 목록 봉투. `total_count` 가 0 이면 `total_pages` 도 0 이다.
      properties:
        page:
          type: integer
          example: 1
        size:
          type: integer
          example: 20
        total_count:
          type: integer
          example: 137
        total_pages:
          type: integer
          example: 7

    # ---------- enum ----------
    UserRole:
      type: string
      enum: [ADMIN, VIEWER]
    UserStatus:
      type: string
      enum: [ACTIVE, SUSPENDED, WITHDRAWN]
...
    UserSummary:
      type: object
      properties:
        user_no:
          type: integer
          format: int64
        user_id:
          type: string
        user_name:
          type: string
          nullable: true
        user_email:
          type: string
          nullable: true
        user_phone:
          type: string
          nullable: true
        user_role:
          $ref: "#/components/schemas/UserRole"
        user_status:
          $ref: "#/components/schemas/UserStatus"
        user_created_at:
          type: string
          format: date-time
          nullable: true
    ActivityType:
      type: string
      description: |
        | 값 | 언제 | `activity_target_no` |
        |---|---|---|
        | `LOGIN` | 로그인 성공 | 없음 |
        | `LOGOUT` | 로그아웃 | 없음 |
        | `PASSWORD_CHANGED` | 비밀번호 변경 성공 | 관리자 변경이면 대상 `user_no` |
        | `PROFILE_UPDATED` | 프로필 수정 성공 | 수정된 `user_no` |
        | `FIRE_CONFIRMED` | 알림에 화재 확인(`READ`) 응답 | `event_no` |
        | `FIRE_DISMISSED` | 알림에 오탐 취소(`CANCEL`) 응답 | `event_no` |

        **실패한 시도는 들어오지 않는다** — 로그인 실패, 거절된 조치 모두 제외다
        (침입 시도·오조작은 활동 이력이 아니다).

        ⚠️ `PUT /api/users/password` 로 바꾼 비밀번호는 아직 `PASSWORD_CHANGED` 가
        남지 않는다. 그 경로만 기록이 빠져 있고 별도로 붙일 예정이다.
      enum: [LOGIN, LOGOUT, PASSWORD_CHANGED, PROFILE_UPDATED,
             FIRE_CONFIRMED, FIRE_DISMISSED]
    UserActivity:
      type: object
      description: |
        사용자 활동이력 1행 (`GET /api/me/activities`,
        `GET /api/users/{user_no}/activities`).
        사용자 1명에 N행이 쌓인다. IP·User-Agent 는 수집하지 않는다.

        `user_no` 는 **행위자**다. 관리자가 남의 계정을 고친 행은 관리자의
        `user_no` 에 쌓이고 `activity_target_no` 가 수정당한 계정을 가리킨다.
      properties:
        activity_no:
          type: integer
          format: int64
        user_no:
          type: integer
          format: int64
          description: 행위자
        activity_type:
          $ref: "#/components/schemas/ActivityType"
        activity_target_no:
          type: integer
          format: int64
          nullable: true
          description: |
            대상 번호. 종류에 따라 `event_no` 이거나 `user_no` 다
            (어느 쪽인지는 `ActivityType` 표 참고). 대상이 없는 종류는 `null`.
        activity_detail:
          type: string
          nullable: true
          description: 화면에 그대로 띄울 한 줄 요약. 없을 수 있다.
        activity_at:
          type: string
          format: date-time
          description: 활동 일시
      example:
        activity_no: 42
        user_no: 1
        activity_type: FIRE_CONFIRMED
        activity_target_no: 12
        activity_detail: 정문 카메라 화재 확인
        activity_at: "2026-08-14T09:12:33"
    UserCreate:
      type: object
      required: [user_id, user_pw, user_name, user_role]
      properties:
        user_id:
          type: string
          pattern: "^[a-z][a-z0-9_-]{4,19}$"
          example: user02
        user_pw:
          type: string
          format: password
          minLength: 8
          maxLength: 64
          example: "Fire#123"
        user_name:
          type: string
          example: 홍길동
        user_role:
          $ref: "#/components/schemas/UserRole"
        user_email:
          type: string
          nullable: true
        user_phone:
          type: string
          nullable: true
          pattern: "^01[016789][0-9]{7,8}$"
          example: "01012345678"
        user_gender:
          type: string
          nullable: true
        user_address:
          type: string
          nullable: true
    UserUpdate:
      type: object
      description: 보낸 필드만 갱신된다. 하나도 없으면 400.
      properties:
        user_name:
          type: string
        user_email:
          type: string
          nullable: true
        user_phone:
          type: string
          nullable: true
          description: 지우려면 `null` 을 보낸다 (빈 문자열은 형식 오류)
          pattern: "^01[016789][0-9]{7,8}$"
        user_pw:
          type: string
          format: password
        user_gender:
          type: string
          nullable: true
        user_address:
          type: string
          nullable: true
        user_role:
          allOf:
            - $ref: "#/components/schemas/UserRole"
          description: ADMIN 만 변경 가능
        user_status:
          allOf:
            - $ref: "#/components/schemas/UserStatus"
          description: ADMIN 만 변경 가능. `WITHDRAWN` 이면 탈퇴 일시가 함께 기록된다

    # ---------- 카메라 ----------
...